Separate tests into different categories
